Add cleanup for manifestworks associated with appSet. Fix App status update by populating a static operate phase (#64)

+++ pkg/controller/multiclusterstatusaggregation/multiclusterStatusAggregation_controller.go
@@ -37,6 +37,7 @@ import (
 	v1 "open-cluster-management.io/api/work/v1"
 	appsetreportV1alpha1 "open-cluster-management.io/multicloud-integrations/pkg/apis/appsetreport/v1alpha1"
 	argov1alpha1 "open-cluster-management.io/multicloud-integrations/pkg/apis/argocd/v1alpha1"
+	propagation "open-cluster-management.io/multicloud-integrations/propagation-controller/application"
 
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
@@ -219,6 +220,10 @@ func (r *ReconcilePullModelAggregation) generateAggregation(ctx context.Context)
 					klog.Warningf("unable to fetch Application: %s", err.Error())
 				} else {
 					newStatus := application.Status.DeepCopy()
+					newStatus.OperationState = &argov1alpha1.OperationState{
+						Phase:     argov1alpha1.OperationRunning,
+						StartedAt: application.CreationTimestamp,
+					}
 					newStatus.Conditions = []argov1alpha1.ApplicationCondition{{
 						Type: "AdditionalStatusReport",
 						Message: fmt.Sprintf("kubectl get multiclusterapplicationsetreports -n %s %s"+
@@ -600,6 +605,55 @@ func (r *ReconcilePullModelAggregation) cleanupOrphanReports() error {
 	}
 
 	for m := range missingAppset {
+		// Look for ManifestWorks associated with this AppSet and then removed them
+		appSet := &missingAppset[m]
+
+		appSetHash, err := propagation.GenerateManifestWorkAppSetHashLabelValue(appSet.Namespace, appSet.Name)
+		if err != nil {
+			klog.Errorf("error generating appSet hash: %v", err)
+			continue
+		}
+
+		appSetRequirement, err := labels.NewRequirement(propagation.LabelKeyAppSetHash, selection.Equals, []string{appSetHash})
+		if err != nil {
+			klog.Errorf("bad requirement: %v", err)
+			continue
+		}
+
+		appSetSelector := labels.NewSelector()
+		appSetSelector = appSetSelector.Add(*appSetRequirement)
+
+		workList := &v1.ManifestWorkList{}
+
+		listopts := &client.ListOptions{
+			LabelSelector: appSetSelector,
+		}
+
+		err = r.List(context.TODO(), workList, listopts)
+		if err != nil {
+			klog.Errorf("Failed to list Argo Application manifestWorks, err: %v", err)
+			continue
+		}
+
+		errOccured := false
+
+		if workList.Items != nil && len(workList.Items) > 0 {
+			for w := range workList.Items {
+				// Remove ManifestWork
+				if err := r.Delete(context.TODO(), &workList.Items[w]); err != nil {
+					if !errors.IsNotFound(err) {
+						klog.Info("Couldn't delete ManifestWork", err)
+
+						errOccured = true
+					}
+				}
+			}
+		}
+
+		if errOccured {
+			continue
+		}
+
 		// Remove orphaned report
 		if err := r.Delete(context.TODO(), &missingAppset[m]); err != nil {
 			if errors.IsNotFound(err) {
